split UTs
